motion_control: add template overload for ControllersIO::set()
Add templated set() method that automatically handles enum and basic
types. Enums are converted to their underlying integer type, eliminating
the need for explicit static_cast when setting values.

Uses ETL's type_traits (etl::is_enum_v) for compile-time type checking
with no runtime overhead or dynamic memory allocation.

Remove unnecessary type casts throughout motion_control filters:
- Remove bool → bool identity casts
- Remove bool → float casts
- Remove enum → float casts

The ParamValue variant supports bool, int, and target_pose_status_t
directly, making explicit casts unnecessary and improving code
readability.

Suggested-by: Mathis Lecrivain <mathis.lecrivain@gmail.com>

+++ motion_control/engines/platform_engine/PlatformEngine.cpp
@@ -54,7 +54,7 @@ void PlatformEngine::prepare_inputs()
     io_.set("angular_target_speed", target_speed_.angle());
 
     // Motion direction
-    io_.set("motion_direction", static_cast<int>(target_pose_.get_motion_direction()));
+    io_.set("motion_direction", target_pose_.get_motion_direction());
 
     // Mark measured values read‑only:
     io_.mark_readonly("motion_direction");
